refactor: update PathEx import paths for consistency across modules
