+++ src/commands/base.ts
@@ -22,7 +22,7 @@ import {type KeyManager} from '../core/key-manager.js';
 import {type AccountManager} from '../core/account-manager.js';
 import {type ProfileManager} from '../core/profile-manager.js';
 import {type CertificateManager} from '../core/certificate-manager.js';
-import {PathEx} from '../core/util/path-ex.js';
+import {PathEx} from '../business/utils/path-ex.js';
 
 export interface Opts {
   logger: SoloLogger;
