+++ src/commands/cluster/tasks.ts
@@ -30,7 +30,7 @@ import {type ClusterRefConnectContext} from './config-interfaces/cluster-ref-con
 import {type ClusterRefDefaultContext} from './config-interfaces/cluster-ref-default-context.js';
 import {type ClusterRefSetupContext} from './config-interfaces/cluster-ref-setup-context.js';
 import {type ClusterRefResetContext} from './config-interfaces/cluster-ref-reset-context.js';
-import {PathEx} from '../../core/util/path-ex.js';
+import {PathEx} from '../../business/utils/path-ex.js';
 
 @injectable()
 export class ClusterCommandTasks {
