+++ src/commands/flags.ts
@@ -19,7 +19,7 @@ import {type AnyListrContext, type AnyObject, type AnyYargs} from '../types/alia
 import {type ClusterRef} from '../core/config/remote/types.js';
 import {type Optional, type SoloListrTaskWrapper} from '../types/index.js';
 import chalk from 'chalk';
-import {PathEx} from '../core/util/path-ex.js';
+import {PathEx} from '../business/utils/path-ex.js';
 
 export class Flags {
   public static KEY_COMMON = '_COMMON_';
