+++ test/e2e/commands/network.test.ts
@@ -19,7 +19,7 @@ import {InitCommand} from '../../../src/commands/init.js';
 import {ClusterCommand} from '../../../src/commands/cluster/index.js';
 import {DeploymentCommand} from '../../../src/commands/deployment.js';
 import {NetworkCommand} from '../../../src/commands/network.js';
-import {PathEx} from '../../../src/core/util/path-ex.js';
+import {PathEx} from '../../../src/business/utils/path-ex.js';
 
 describe('NetworkCommand', function networkCommand() {
   this.bail(true);
