+++ test/test-container.ts
@@ -6,7 +6,7 @@ import {type NamespaceNameAsString} from '../src/core/config/remote/types.js';
 import * as yaml from 'yaml';
 import {DEFAULT_LOCAL_CONFIG_FILE} from '../src/core/constants.js';
 import {type SoloLogger} from '../src/core/logging.js';
-import {PathEx} from '../src/core/util/path-ex.js';
+import {PathEx} from '../src/business/utils/path-ex.js';
 
 const cacheDirectory = PathEx.join('test', 'data', 'tmp');
