+++ test/unit/commands/cluster.test.ts
@@ -23,7 +23,7 @@ import {DependencyManager} from '../../../src/core/dependency-managers/index.js'
 import {NamespaceName} from '../../../src/core/kube/resources/namespace/namespace-name.js';
 import {InjectTokens} from '../../../src/core/dependency-injection/inject-tokens.js';
 import {Argv} from '../../helpers/argv-wrapper.js';
-import {PathEx} from '../../../src/core/util/path-ex.js';
+import {PathEx} from '../../../src/business/utils/path-ex.js';
 
 const getBaseCommandOpts = (context: string) => {
   const opts = {
