+++ test/unit/core/zippy.test.ts
@@ -12,7 +12,7 @@ import os from 'os';
 import fs from 'fs';
 import {Zippy} from '../../../src/core/zippy.js';
 import * as logging from '../../../src/core/logging.js';
-import {PathEx} from '../../../src/core/util/path-ex.js';
+import {PathEx} from '../../../src/business/utils/path-ex.js';
 
 describe('Zippy', () => {
   const testLogger = logging.NewLogger('debug', true);
